Added code-style and formatted existing code (#18)
closes #15
