--- src/main/java/ca/softwarespace/qiyanna/dataaggregator/Utils/Constants.java
@@ -1,10 +1,0 @@
-package ca.softwarespace.qiyanna.dataaggregator.Utils;
-
-/**
- * Author: Steve Mbiele
- * Date: 5/15/2019
- */
-public class Constants {
-    public static final String MATCH_LIST_BEGIN_INDEX_PARAMETER = "beginIndex";
-    public static final String MATCHES_NODE_IN_MATCH_LIST = "matches";
-}
