+++ src/main/java/ca/softwarespace/qiyanna/dataaggregator/util/Constants.java
@@ -1,0 +1,10 @@
+package ca.softwarespace.qiyanna.dataaggregator.util;
+
+/**
+ * Author: Steve Mbiele Date: 5/15/2019
+ */
+public class Constants {
+
+  public static final String MATCH_LIST_BEGIN_INDEX_PARAMETER = "beginIndex";
+  public static final String MATCHES_NODE_IN_MATCH_LIST = "matches";
+}
